+++ odoo_rpc_client/connection/__init__.py
@@ -1,6 +1,5 @@
 from . import (xmlrpc,   # noqa
-               jsonrpc,  # noqa
-               local)    # noqa
+               jsonrpc)  # noqa
 from .connection import (ConnectorBase,        # noqa
                          get_connector,        # noqa
                          get_connector_names)  # noqa