+++ odoo_rpc_client/tests/__init__.py
@@ -2,7 +2,6 @@ import os
 import six
 import unittest
 from ..utils import AttrDict
-from ..connection.local import ConnectorLocal
 
 import logging
 _logger = logging.getLogger(__name__)
@@ -16,17 +15,6 @@ except ImportError:
     import mock
 
 __all__ = ('BaseTestCase', 'mock')
-
-
-class ConnectorLocalTestExt(ConnectorLocal):
-    """ Extend local connector to get odoo options from environment
-    """
-
-    def _start_odoo_services(self):
-        if not self.odoo_args:
-            self.odoo_args = eval(
-                os.environ.get('ODOO_TEST_LOCAL_ARGS', '[]'))
-        return super(ConnectorLocalTestExt, self)._start_odoo_services()
 
 
 class BaseTestCase(unittest.TestCase):
@@ -47,7 +35,6 @@ class BaseTestCase(unittest.TestCase):
             'password': os.environ.get('ODOO_TEST_PASSWORD', 'admin'),
             'super_password': os.environ.get('ODOO_TEST_SUPER_PASSWORD',
                                              'admin'),
-            'local_args': eval(os.environ.get('ODOO_TEST_LOCAL_ARGS', 'None'))
         })
 
         self.test_db_service = os.environ.get('TEST_DB_SERVICE', False)