+++ CHANGELOG.rst
@@ -2,13 +2,16 @@ Last changes
 ============
 
 
-Release 0.8.3
+Release 0.9.0
 -------------
 
 - Removed old alias *odoo_rpc_client.orm.record.RecordRelations
 - Added *search_count* method to objects (models)
 - Added *create_record* method to objects (models)
 - Added tests for Odoo 11
+- Removed *local* database connector, because it is difficult to test it,
+  and it is not used so often as *xml-rpc* and *json-rpc* connectors
+  May be in future it will have it's own python package.
 
 
 Release 0.8.2